+++ src/lib.rs
@@ -7,5 +7,6 @@ pub mod common;
 pub mod ipc;
 pub mod layout_engine;
 pub mod model;
+pub mod status;
 pub mod sys;
 pub mod ui;